refactor: 데드코드 제거 및 내부 컴포넌트 폴더 구조 정리
- BoardEmpty, CourseRow, DayBar를 MenuBoard 전용 내부 폴더(_*)로 이동
- board/index.ts에서 내부 컴포넌트 export 제거, MenuBoard만 공개
- @shared/index.ts를 sub-barrel 방식(export * from './error|layout')으로 전환
- src/components/index.ts, vitest.config.ts 삭제 (미사용)
- PickType 'A'|'B'|'takeout' → 'COURSE_1'|'COURSE_2'|'TAKE_OUT'으로 통일
- CATEGORY_TO_PICK 브릿지 상수 제거
- CAFETERIA export 제거 (내부 전용)
- 테스트 관련 패키지 제거 (vitest, @testing-library/*, jsdom, vite)
- cleanup route date 포맷 버그 수정 (YYYYMMDD → YYYY-MM-DD)
