+++ package.json
@@ -10,9 +10,7 @@
     "lint:fix": "next lint --fix",
     "format": "prettier --write .",
     "format:check": "prettier --check .",
-    "mcp": "npx tsx src/mcp/index.ts",
-    "test": "vitest run",
-    "test:watch": "vitest"
+    "mcp": "npx tsx src/mcp/index.ts"
   },
   "dependencies": {
     "@modelcontextprotocol/sdk": "^1.27.1",
@@ -34,21 +32,15 @@
   "devDependencies": {
     "@eslint/eslintrc": "^3",
     "@next/eslint-plugin-next": "^15.4.2",
-    "@testing-library/jest-dom": "^6.9.1",
-    "@testing-library/react": "^16.3.2",
     "@types/node": "^20",
     "@types/react": "^19",
     "@types/react-dom": "^19",
-    "@vitejs/plugin-react": "^4.7.0",
     "eslint": "^9",
     "eslint-config-next": "15.3.5",
     "eslint-plugin-react-hooks": "^5.2.0",
-    "jsdom": "^24.1.3",
     "prettier": "^3.6.2",
     "prettier-plugin-tailwindcss": "^0.6.14",
     "tsx": "^4.21.0",
-    "typescript": "^5",
-    "vite": "^5.4.21",
-    "vitest": "^2.1.9"
+    "typescript": "^5"
   }
 }
