+++ src/components/@shared/index.ts
@@ -1,3 +1,2 @@
-export {default as ErrorBoundary} from './error/ErrorBoundary';
-export {default as SiteFooter} from './layout/SiteFooter';
-export {default as SiteHeader} from './layout/SiteHeader';
+export * from './error';
+export * from './layout';
