--- src/components/board/BoardEmpty/index.ts
@@ -1,1 +1,0 @@
-export {default} from './BoardEmpty';
