--- src/components/board/CourseRow/index.ts
@@ -1,1 +1,0 @@
-export {default} from './CourseRow';
