--- src/components/board/DayBar/index.ts
@@ -1,1 +1,0 @@
-export {default} from './DayBar';
