+++ src/components/board/MenuBoard/_BoardEmpty/BoardEmpty.tsx
