+++ src/components/board/MenuBoard/_BoardEmpty/BoardEmpty.types.ts
