+++ src/components/board/MenuBoard/_BoardEmpty/constants.ts
