+++ src/components/board/MenuBoard/_CourseRow/CourseRow.styles.ts
