+++ src/components/board/MenuBoard/_CourseRow/CourseRow.tsx
