+++ src/components/board/MenuBoard/_CourseRow/CourseRow.types.ts
