+++ src/components/board/MenuBoard/_DayBar/DayBar.types.ts
