+++ src/components/board/MenuBoard/_DayBar/DayChip.styles.ts
