+++ src/components/board/index.ts
@@ -1,4 +1,1 @@
-export {default as BoardEmpty} from './BoardEmpty';
-export {default as CourseRow} from './CourseRow';
-export {default as DayBar} from './DayBar';
 export {default as MenuBoard} from './MenuBoard';
