--- src/components/index.ts
@@ -1,3 +1,0 @@
-export * from './board';
-export * from './home';
-export * from './@shared';
