+++ src/lib/menu-policy.ts
@@ -4,7 +4,7 @@ import {getWeekDays} from '@/lib/utils';
 const pad = (n: number) => String(n).padStart(2, '0');
 
 /** 구내식당 운영 시각 config — 여기만 수정하면 전체 반영 */
-export const CAFETERIA = {
+const CAFETERIA = {
   openHour: 11,
   openMinute: 0,
   closeHour: 13,
