--- vitest.config.ts
@@ -1,16 +1,0 @@
-import path from 'path';
-
-import react from '@vitejs/plugin-react';
-import {defineConfig} from 'vitest/config';
-
-export default defineConfig({
-  plugins: [react()],
-  test: {
-    environment: 'jsdom',
-    setupFiles: ['./tests/setup.ts'],
-    include: ['tests/**/*.test.{ts,tsx}'],
-  },
-  resolve: {
-    alias: {'@': path.resolve(__dirname, './src')},
-  },
-});
